move propsDecoder (#41)
* move propsDecoder

* Add Request as type

* Update workers-types

* Fix lint error
