+++ evm-gateway/src/index.ts
@@ -1,4 +1,4 @@
 export { EVMGateway, StorageLayout } from './EVMGateway.js';
 export { EVMProofHelper, type StateProof } from './EVMProofHelper.js';
 export type { IProofService, ProvableBlock } from './IProofService.js';
-export type { Router } from './utils.js';
+export { propsDecoder, type Router } from './utils.js';
